Basic Monsters Being Rendered
Once again... my GitHub had stopped working. All of the changes were
being synced on this client but not in my directory. When I made the
test project it copied the contents of the github project into my
workspace, so when I made changes it was on a different copy. I will be
trying to put my changes into this now that I have it fixed. I will
start to tidy up all the monster stuff and get all the functions working
together. Sorry about that. I was getting suspicious that something was
wrong with my copy as changes didn't seem to be occurring, but as they
were listed on this, I thought they were being synced and all was well.